bot/requirements.txt
# Requires Python 3.10+ (claude-agent-sdk constraint)
aiogram>=3.15
claude-agent-sdk>=0.1.44
python-dotenv>=1.0
aiosqlite>=0.20
